requirements.txt
openai
faiss-cpu
pyyaml
numpy
sentence-transformers
transformers
